rtio: Split header into logical components
Much like the kernel or llext, rtio.h had become quite large though the
concepts and data types are not.

Restructure the headers into logically grouped components.
